article-script/requirements.txt
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
et-xmlfile==1.1.0
google-api-core==1.31.1
google-api-python-client==1.12.8
google-auth==1.34.0
google-auth-httplib2==0.1.0
Google-Images-Search==1.3.8
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
numpy==1.21.1
openai==0.10.2
openpyxl==3.0.7
packaging==21.0
pandas==1.3.1
pandas-stubs==1.2.0.1
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyfiglet==0.8.post1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-resize-image==1.1.19
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.62.0
uritemplate==3.0.1
urllib3==1.26.6
